Clean-up and make importable:
- rename to btor2opt
- add __init__.py
- refactor abstract pass class from instantiations
