+++ __init__.py
@@ -1,0 +1,5 @@
+
+
+from .src import program
+from .src.parse import parse
+from .src.passes.passes import Pass
