+++ src/passes/transforms/renameInputs.py
@@ -1,0 +1,17 @@
+# Example pass: Simply renames all inputs to inp_<pos>
+class RenameInputs(Pass):
+    def __init__(self):
+        super().__init__("rename-inputs")
+
+    # I chose to have this pass not modify p in place
+    # you can also simply modify p and return it
+    def run(p: list[Instruction]) -> list[Instruction]:
+        i = 0
+        res = []
+        for inst in p:
+            if isinstance(inst, Input):
+                res.append(Input(inst.lid, inst.sort, f"inp_{i}"))
+                i += 1
+            else:
+                res.append(inst)
+        return res
